external cav setter is moved to a separate mixin
