+++ src/ggrc/models/mixins/with_readonly_access.py
@@ -31,7 +31,6 @@ class WithReadOnlyAccess(object):
 
   _api_attrs = reflection.ApiAttributes(
       reflection.Attribute("readonly", create=True, update=True),
-      reflection.Attribute("external_custom_attributes", read=False),
   )
 
   _aliases = {
@@ -79,54 +78,3 @@ class WithReadOnlyAccess(object):
       raise ValidationError("Attribute 'readonly' has invalid value")
 
     return value
-
-  def external_custom_attributes(self, src):
-    # type: (List[Dict[str, str]]) -> None
-    """Set CAV in format specific to sync service
-
-    This method accesses only 'external_custom_attributes' key from 'src'
-
-    Format example: [{'name': 'a', value: 'val-a'},
-                     {'name': 'b', value: 'val-b'}]
-    In this example, values 'val-a' and 'val-b' have to be set to
-    custom attributes with names 'a' and 'b' correspondingly
-
-    if `values` is empty, no attributes will be set
-
-    If custom attribute with some name doesn't exist, corresponding item will
-    be skipped
-
-    This method calls original `custom_attribute_values` setter
-
-    :param src: json dict from request
-    """
-
-    values = src.get('external_custom_attributes')
-
-    if not values:
-      return
-
-    cad_ti_map = dict(
-        (cad.title, cad.id)
-        for cad in self.get_custom_attribute_definitions()
-    )
-
-    cavs = list()
-    for item in values:
-      if 'name' not in item or 'value' not in item:
-        continue
-
-      name = item['name']
-      value = item['value']
-      if name not in cad_ti_map:
-        logger.warning("Skipping external custom attribute %r, because "
-                       "definition with such name was not found", name)
-        continue
-      cavs.append({
-          'custom_attribute_id': cad_ti_map[name],
-          'attribute_value': value,
-          'attribute_object': None,
-      })
-
-    # pylint: disable=attribute-defined-outside-init
-    self.custom_attribute_values = cavs
