+++ src/ggrc/models/system.py
@@ -11,6 +11,7 @@ from ggrc.models.comment import ScopedCommentable
 from ggrc.models.deferred import deferred
 from ggrc.models import mixins
 from ggrc.models.mixins import synchronizable
+from ggrc.models.mixins.with_ext_custom_attrs import WithExtCustomAttrsSetter
 from ggrc.models.mixins.with_readonly_access import WithReadOnlyAccess
 from ggrc.models.object_document import PublicDocumentable
 from ggrc.models.object_person import Personable
@@ -79,7 +80,7 @@ class SystemOrProcess(ScopedCommentable,
     )
 
 
-class System(mixins.CustomAttributable,
+class System(WithExtCustomAttrsSetter,
              WithReadOnlyAccess,
              Personable,
              synchronizable.RoleableSynchronizable,
